requirements.txt
jupyter===1.0.0
ipykernel===6.29.4
matplotlib
numpy
pytest===7.4.3
coverage===7.4.4
flake8===7.0.0
pandas===2.2.2
plotly
replicate
vertexai===1.48.0
streamlit===1.35.0
streamlit-feedback
langchain
streamlit-chat===0.1.1
streamlit-extras===0.4.2
streamlit-embedcode
barfi
notion-client===2.2.1
opencv-python===4.9.0.80
pytesseract===0.3.10
deskew===1.5.1
streamlit-chat===0.1.1
streamlit-pdf-viewer===0.0.10
openpyxl===3.1.2
google-generativeai @ git+https://github.com/joe-stifler/generative-ai-python@d291b29d01d257e9d7079ef30d7bd43670b1a4ec
pdf2image===1.17.0
tabulate===0.9.0
pydantic===2.7.1
notion-client===2.2.1
llama-index===0.10.36
python-dotenv===1.0.1
black[jupyter]===24.4.2
PyMuPDF===1.24.4
